langgraph_agent_app_sample_code/requirements.txt
pydantic>=2.9.2
databricks-agents
mlflow>=2.18.0
databricks-sdk[openai]
databricks-vectorsearch
pyyaml
langgraph
langchain_core
databricks-langchain
langchain-community 
git+https://github.com/unitycatalog/unitycatalog.git#subdirectory=ai/integrations/langchain
git+https://github.com/unitycatalog/unitycatalog.git#subdirectory=ai/core
git+https://github.com/unitycatalog/unitycatalog.git#subdirectory=ai/integrations/openai
tabulate
pandas
pyspark
databricks-connect==15.1.0
pytest
